+++ Assets/Scripts/Enemies/SwarmUnit.cs
@@ -64,7 +64,7 @@ namespace usd.Enemies
         
         // Override take damage and die methods
         
-        public new void TakeDamage(int damageTaken)
+        public new void TakeDamage(float damageTaken)
         {
             health -= damageTaken;
             if (health <= 0)
